// src/model/Lipei_view.ts
export class Lipei_view{
    FramGuid: string;
    FramType: string;
    GPS: string;
    insuranceId: string;
    name: string;
    idCard: string;
    bankName: string;
    bankAccount: string;
    accountType: object;
    startTime: string;
    endTime: string;
    BanNum: string;
    reportDate: string;
    outDate: string;
    contactNumber: string;
    address: string;
    FramIDFont: string;
    FramIDBack: string;
    VaccinationCertificateImg: string;
    InstitutionCodeImg: string;
    BankCardImg: string;
    remarks: string;
    CreateTime: string;
    vaccinationCertificateImgs: string;

    constructor() {
        this.FramGuid = "";
        this.FramType = "";
        this.GPS = "";
        this.insuranceId = "";
        this.name = "";
        this.idCard = "";
        this.bankName = "";
        this.bankAccount = "";
        this.accountType = {};
        this.startTime = "";
        this.endTime = "";
        this.BanNum = "";
        this.reportDate = "";
        this.outDate = "";
        this.contactNumber = "";
        this.address = "";
        this.FramIDFont = "";
        this.FramIDBack = "";
        this.VaccinationCertificateImg = "";
        this.InstitutionCodeImg = "";
        this.BankCardImg = "";
        this.remarks = "";
        this.CreateTime = "";
        this.vaccinationCertificateImgs = "";
    }
}